refactor: move external-auth-mapping to data folder
