+++ src/infra/db/prisma/data/external-auth-mapping/find-external-auth-mapping/find-external-auth-mapping-by-external-auth-user-id-prisma-repo.spec.ts
